# internal/goa/service/resume/gen/http/hy_user_work_history/server/encode_decode.go
// Code generated by goa v3.7.13, DO NOT EDIT.
//
// hy_userWorkHistory HTTP server encoders and decoders
//
// Command:
// $ goa gen resume/design

package server

import (
	"context"
	"net/http"
	hyuserworkhistoryviews "resume/gen/hy_user_work_history/views"
	"strconv"
	"strings"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetUserWorkHistoryResponse returns an encoder for responses returned
// by the hy_userWorkHistory getUserWorkHistory endpoint.
func EncodeGetUserWorkHistoryResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(hyuserworkhistoryviews.UserworkhistoryCollection)
		enc := encoder(ctx, w)
		body := NewUserworkhistoryResponseCollection(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetUserWorkHistoryRequest returns a decoder for requests sent to the
// hy_userWorkHistory getUserWorkHistory endpoint.
func DecodeGetUserWorkHistoryRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			userID int
			token  *string
			err    error

			params = mux.Vars(r)
		)
		{
			userIDRaw := params["userID"]
			v, err2 := strconv.ParseInt(userIDRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("userID", userIDRaw, "integer"))
			}
			userID = int(v)
		}
		if userID < 1 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("userID", userID, 1, true))
		}
		tokenRaw := r.Header.Get("Authorization")
		if tokenRaw != "" {
			token = &tokenRaw
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetUserWorkHistoryPayload(userID, token)
		if payload.Token != nil {
			if strings.Contains(*payload.Token, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.Token, " ", 2)[1]
				payload.Token = &cred
			}
		}

		return payload, nil
	}
}

// marshalHyuserworkhistoryviewsUserworkhistoryViewToUserworkhistoryResponse
// builds a value of type *UserworkhistoryResponse from a value of type
// *hyuserworkhistoryviews.UserworkhistoryView.
func marshalHyuserworkhistoryviewsUserworkhistoryViewToUserworkhistoryResponse(v *hyuserworkhistoryviews.UserworkhistoryView) *UserworkhistoryResponse {
	res := &UserworkhistoryResponse{
		Title:       *v.Title,
		Company:     *v.Company,
		Country:     *v.Country,
		Term:        v.Term,
		Description: v.Description,
		Techs:       v.Techs,
	}

	return res
}