# internal/goa/service/resume/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: resume API
    description: go-goa API with goa framework
    contact:
        name: hiromaily
        url: https://hiromaily.github.io/
        email: hiromaily@gmail.com
    license:
        name: MIT
        url: https://github.com/goadesign/goa/blob/master/LICENSE
    version: "1.0"
servers:
    - url: http://localhost:8080/api
paths:
...
    /user:
        get:
            tags:
                - hy_user
            summary: userList hy_user
            description: List all users
            operationId: hy_user#userList
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                description: 'Response body may alternatively be #/components/schemas/UserResponseIDCollection'
                                $ref: '#/components/schemas/UserCollection'
                            example:
                                - created_at: "2017-03-10T15:00:00Z"
                                  email: hy@gmail.com
                                  id: 10
                                  password: xxxxxxxx
                                  updated_at: "2017-03-10T15:00:00Z"
                                  user_name: Hiroki
                                - created_at: "2017-03-10T15:00:00Z"
                                  email: hy@gmail.com
                                  id: 10
                                  password: xxxxxxxx
                                  updated_at: "2017-03-10T15:00:00Z"
                                  user_name: Hiroki
                                - created_at: "2017-03-10T15:00:00Z"
                                  email: hy@gmail.com
                                  id: 10
                                  password: xxxxxxxx
                                  updated_at: "2017-03-10T15:00:00Z"
                                  user_name: Hiroki
                                - created_at: "2017-03-10T15:00:00Z"
                                  email: hy@gmail.com
                                  id: 10
                                  password: xxxxxxxx
                                  updated_at: "2017-03-10T15:00:00Z"
                                  user_name: Hiroki
            security:
                - jwt_header_Authorization:
                    - api:access
        post:
            tags:
                - hy_user
            summary: createUser hy_user
            description: Create new user
            operationId: hy_user#createUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequestBody'
                        example:
                            email: hy@gmail.com
                            password: xxxxxxxx
                            user_name: Hiroki
                            userID: 6881235799436279149
            responses:
                "200":
                    description: OK response.
                "201":
                    description: Created response.
            security:
                - jwt_header_Authorization:
                    - api:access
    /user/{userID}:
        delete:
            tags:
                - hy_user
            summary: deleteUser hy_user
            description: Delete user
            operationId: hy_user#deleteUser
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 1392782480822914393
                    minimum: 1
                  example: 5692038594719316570
            responses:
                "200":
                    description: OK response.
            security:
                - jwt_header_Authorization:
                    - api:access
        get:
            tags:
                - hy_user
            summary: getUser hy_user
            description: get user with given user id
            operationId: hy_user#getUser
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 4961098857206787529
                    minimum: 1
                  example: 2289587835042610118
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                description: 'Response body may alternatively be #/components/schemas/UserResponseID'
                                $ref: '#/components/schemas/User'
                            example:
                                created_at: "2017-03-10T15:00:00Z"
                                email: hy@gmail.com
                                id: 10
                                password: xxxxxxxx
                                updated_at: "2017-03-10T15:00:00Z"
                                user_name: Hiroki
            security:
                - jwt_header_Authorization:
                    - api:access
        put:
            tags:
                - hy_user
            summary: updateUser hy_user
            description: Change user properties
            operationId: hy_user#updateUser
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 7506736133595476514
                    minimum: 1
                  example: 1748901102406723351
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserRequestBody'
                        example:
                            email: hy@gmail.com
                            password: xxxxxxxx
                            user_name: Hiroki
            responses:
                "200":
                    description: OK response.
            security:
                - jwt_header_Authorization:
                    - api:access
...
    /user/{userID}/workhistory:
        get:
            tags:
                - hy_userWorkHistory
            summary: getUserWorkHistory hy_userWorkHistory
            description: get user's work history
            operationId: hy_userWorkHistory#getUserWorkHistory
            parameters:
                - name: userID
                  in: path
                  description: User ID
                  required: true
                  schema:
                    type: integer
                    description: User ID
                    example: 4284632955958688991
                    minimum: 1
                  example: 2846710434179190158
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserworkhistoryCollection'
                            example:
                                - company: Company
                                  country: nl
                                  description: '["Developed resume site for job seeking."]'
                                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                                  term: Jul 2017 - Aug 2017
                                  title: Developer
                                - company: Company
                                  country: nl
                                  description: '["Developed resume site for job seeking."]'
                                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                                  term: Jul 2017 - Aug 2017
                                  title: Developer
            security:
                - jwt_header_Authorization:
                    - api:access
components:
    schemas:
...
        CreateUserRequestBody:
            type: object
            properties:
                email:
                    type: string
                    description: E-mail of user
                    example: hy@gmail.com
                    format: email
                password:
                    type: string
                    description: Password
                    example: xxxxxxxx
                    minLength: 8
                    maxLength: 20
                user_name:
                    type: string
                    description: User name
                    example: Hiroki
                    minLength: 2
                    maxLength: 20
                userID:
                    type: integer
                    description: User ID
                    example: 2919063293485881591
                    minimum: 1
            example:
                email: hy@gmail.com
                password: xxxxxxxx
                user_name: Hiroki
                userID: 6333285541576646153
            required:
                - user_name
                - email
                - password
...
        UpdateUserRequestBody:
            type: object
            properties:
                email:
                    type: string
                    description: E-mail of user
                    example: hy@gmail.com
                    format: email
                password:
                    type: string
                    description: Password
                    example: xxxxxxxx
                    minLength: 8
                    maxLength: 20
                user_name:
                    type: string
                    description: User name
                    example: Hiroki
                    minLength: 2
                    maxLength: 20
            example:
                email: hy@gmail.com
                password: xxxxxxxx
                user_name: Hiroki
            required:
                - user_name
                - email
                - password
        User:
            type: object
            properties:
                created_at:
                    type: string
                    description: Datetime
                    example: "2017-03-10T15:00:00Z"
                    format: date-time
                email:
                    type: string
                    description: E-mail of user
                    example: hy@gmail.com
                    format: email
                id:
                    type: integer
                    description: ID
                    example: 10
                    minimum: 1
                password:
                    type: string
                    description: Password
                    example: xxxxxxxx
                    minLength: 8
                    maxLength: 20
                updated_at:
                    type: string
                    description: Datetime
                    example: "2017-03-10T15:00:00Z"
                    format: date-time
                user_name:
                    type: string
                    description: User name
                    example: Hiroki
                    minLength: 2
                    maxLength: 20
            description: A user information
            example:
                created_at: "2017-03-10T15:00:00Z"
                email: hy@gmail.com
                id: 10
                password: xxxxxxxx
                updated_at: "2017-03-10T15:00:00Z"
                user_name: Hiroki
            required:
                - user_name
                - email
                - password
        UserCollection:
            type: array
            items:
                $ref: '#/components/schemas/User'
            example:
                - created_at: "2017-03-10T15:00:00Z"
                  email: hy@gmail.com
                  id: 10
                  password: xxxxxxxx
                  updated_at: "2017-03-10T15:00:00Z"
                  user_name: Hiroki
                - created_at: "2017-03-10T15:00:00Z"
                  email: hy@gmail.com
                  id: 10
                  password: xxxxxxxx
                  updated_at: "2017-03-10T15:00:00Z"
                  user_name: Hiroki
        UserResponseID:
            type: object
            properties:
                id:
                    type: integer
                    description: ID
                    example: 10
                    minimum: 1
            description: A user information (id view)
            example:
                id: 10
        UserResponseIDCollection:
            type: array
            items:
                $ref: '#/components/schemas/UserResponseID'
            description: UserListResponseBody is the result type for an array of UserResponse (id view)
            example:
                - id: 10
                - id: 10
                - id: 10
                - id: 10
...
        Userworkhistory:
            type: object
            properties:
                company:
                    type: string
                    description: Company name
                    example: Company
                    minLength: 2
                    maxLength: 40
                country:
                    type: string
                    description: Country code
                    example: nl
                    minLength: 2
                    maxLength: 2
                description:
                    type: string
                    description: job description
                    example: '["Developed resume site for job seeking."]'
                    format: binary
                techs:
                    type: string
                    description: used techs
                    example: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                    format: binary
                term:
                    type: string
                    description: worked period
                    example: Jul 2017 - Aug 2017
                    minLength: 10
                    maxLength: 20
                title:
                    type: string
                    description: Job Title
                    example: Developer
                    minLength: 2
                    maxLength: 40
            description: A user information
            example:
                company: Company
                country: nl
                description: '["Developed resume site for job seeking."]'
                techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                term: Jul 2017 - Aug 2017
                title: Developer
            required:
                - title
                - company
                - country
        UserworkhistoryCollection:
            type: array
            items:
                $ref: '#/components/schemas/Userworkhistory'
            example:
                - company: Company
                  country: nl
                  description: '["Developed resume site for job seeking."]'
                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                  term: Jul 2017 - Aug 2017
                  title: Developer
                - company: Company
                  country: nl
                  description: '["Developed resume site for job seeking."]'
                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                  term: Jul 2017 - Aug 2017
                  title: Developer
                - company: Company
                  country: nl
                  description: '["Developed resume site for job seeking."]'
                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                  term: Jul 2017 - Aug 2017
                  title: Developer
                - company: Company
                  country: nl
                  description: '["Developed resume site for job seeking."]'
                  techs: '["Golang with goa", "Riot.js", "Semantic UI", "MySQL", "Docker", "Travis-CI"]'
                  term: Jul 2017 - Aug 2017
                  title: Developer
    securitySchemes:
        jwt_header_Authorization:
            type: http
            scheme: bearer
...